Abstract Claim Rewards PTB
