+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "navi-sdk",
-  "version": "1.2.1",
+  "version": "1.2.2",
   "description": "",
   "main": "dist/index.js",
   "scripts": {
