+++ src/libs/PTB/commonFunctions.ts
@@ -2,6 +2,15 @@ import { Transaction } from "@mysten/sui/transactions";
 import { getConfig, flashloanConfig, pool, vSuiConfig } from '../../address'
 import { CoinInfo, Pool, PoolConfig, OptionType } from '../../types';
 import { bcs } from "@mysten/sui/bcs";
+import { getFullnodeUrl, SuiClient } from "@mysten/sui/client";
+import { moveInspect } from "../CallFunctions";
+
+interface Reward {
+    asset_id: string;
+    funds: string;
+    available: string;
+  }
+
 /**
  * Deposits a specified amount of a coin into a pool.
  * @param txb - The transaction block object.
@@ -119,7 +128,7 @@ export async function withdrawCoinWithAccountCap(txb: Transaction, _pool: PoolCo
                 objectId: '0x06',
                 initialSharedVersion: 1,
                 mutable: false,
-              }), // clock object id
+            }), // clock object id
             txb.object(config.PriceOracle), // object id of oracle
             txb.object(config.StorageId), // object id of storage
             txb.object(_pool.poolId), // pool id of the asset
@@ -222,7 +231,7 @@ export async function getHealthFactor(txb: Transaction, address: string) {
             txb.object(config.PriceOracle), // Object id of Price Oracle
             txb.pure.address(address)
         ],
-        
+
     })
 
     return balance;
@@ -247,12 +256,12 @@ export function returnMergedCoins(txb: Transaction, coinInfo: any) {
     // }
 
     if (coinInfo.data.length >= 2) {
-            let baseObj = coinInfo.data[0].coinObjectId;
-            let all_list = coinInfo.data.slice(1).map((coin: any) => coin.coinObjectId);
+        let baseObj = coinInfo.data[0].coinObjectId;
+        let all_list = coinInfo.data.slice(1).map((coin: any) => coin.coinObjectId);
 
-            txb.mergeCoins(baseObj, all_list);
-        }
-    
+        txb.mergeCoins(baseObj, all_list);
+    }
+
     let mergedCoinObject = txb.object(coinInfo.data[0].coinObjectId);
     return mergedCoinObject;
 }
@@ -355,22 +364,22 @@ export async function claimRewardFunction(txb: Transaction, incentiveFundsPool: 
 
     const ProFundsPoolInfo: any = {
         'f975bc2d4cca10e3ace8887e20afd77b46c383b4465eac694c4688344955dea4': {
-          coinType: '0x2::sui::SUI',
-          oracleId: 0,
+            coinType: '0x2::sui::SUI',
+            oracleId: 0,
         },
         'e2b5ada45273676e0da8ae10f8fe079a7cec3d0f59187d3d20b1549c275b07ea': {
-          coinType: '0x549e8b69270defbfafd4f94e17ec44cdbdd99820b33bda2278dea3b9a32d3f55::cert::CERT',
-          oracleId: 5,
+            coinType: '0x549e8b69270defbfafd4f94e17ec44cdbdd99820b33bda2278dea3b9a32d3f55::cert::CERT',
+            oracleId: 5,
         },
         'a20e18085ce04be8aa722fbe85423f1ad6b1ae3b1be81ffac00a30f1d6d6ab51': {
-          coinType: '0xbde4ba4c2e274a60ce15c1cfff9e5c42e41654ac8b6d906a57efa4bd3c29f47d::hasui::HASUI',
-          oracleId: 6,
+            coinType: '0xbde4ba4c2e274a60ce15c1cfff9e5c42e41654ac8b6d906a57efa4bd3c29f47d::hasui::HASUI',
+            oracleId: 6,
         },
         '9dae0cf104a193217904f88a48ce2cf0221e8cd9073878edd05101d6b771fa09': {
-          coinType: '0xa99b8952d4f7d947ea77fe0ecdcc9e5fc0bcab2841d6e2a5aa00c3044e5544b5::navx::NAVX',
-          oracleId: 7,
+            coinType: '0xa99b8952d4f7d947ea77fe0ecdcc9e5fc0bcab2841d6e2a5aa00c3044e5544b5::navx::NAVX',
+            oracleId: 7,
         },
-      }
+    }
     txb.moveCall({
         target: `${config.ProtocolPackage}::incentive_v2::claim_reward`,
         arguments: [
@@ -412,7 +421,7 @@ export async function SignAndSubmitTXB(txb: Transaction, client: any, keypair: a
  * @returns vSui coin object.
  */
 export async function stakeTovSui(txb: Transaction, suiCoinObj: any) {
-    
+
     const [coin] = txb.moveCall({
         target: `${vSuiConfig.ProtocolPackage}::native_pool::stake_non_entry`,
         arguments: [
@@ -433,7 +442,7 @@ export async function stakeTovSui(txb: Transaction, suiCoinObj: any) {
  * @returns The unstaked Sui coin.
  */
 export async function unstakeTovSui(txb: Transaction, vSuiCoinObj: any) {
-    
+
     const [coin] = txb.moveCall({
         target: `${vSuiConfig.ProtocolPackage}::native_pool::unstake`,
         arguments: [
@@ -446,3 +455,117 @@ export async function unstakeTovSui(txb: Transaction, vSuiCoinObj: any) {
     })
     return coin;
 }
+
+  /**
+   * Retrieves the incentive pools for a given asset and option.
+   * @param assetId - The ID of the asset.
+   * @param option - The option type.
+   * @param user - (Optional) The user's address. If provided, the rewards claimed by the user and the total rewards will be returned.
+   * @returns The incentive pools information.
+   */
+export async function getIncentivePools(client: SuiClient, assetId: number, option: OptionType, user:string) {
+    const config = await getConfig();
+    const tx = new Transaction();
+    const result: any = await moveInspect(
+        tx,
+        client,
+        user,
+        `${config.uiGetter}::incentive_getter::get_incentive_pools`,
+        [
+            tx.object('0x06'), // clock object id
+            tx.object(config.IncentiveV2), // the incentive object v2
+            tx.object(config.StorageId), // object id of storage
+            tx.pure.u8(assetId),
+            tx.pure.u8(option),
+            tx.pure.address(user), // If you provide your address, the rewards that have been claimed by your address and the total rewards will be returned.
+        ],
+        [], // type arguments is null
+        'vector<IncentivePoolInfo>' // parse type
+    );
+    return result[0];
+}
+
+  /**
+   * Retrieves the available rewards for a given address.
+   * 
+   * @param checkAddress - The address to check for rewards. Defaults to the current address.
+   * @param option - The option type. Defaults to 1.
+   * @param prettyPrint - Whether to print the rewards in a pretty format. Defaults to true.
+   * @returns An object containing the summed rewards for each asset.
+   * @throws If there is an error retrieving the available rewards.
+   */
+export async function getAvailableRewards(client: SuiClient, checkAddress: string, option: OptionType = 1, prettyPrint = true) {
+    const assetIds = Array.from({ length: 8 }, (_, i) => i); // Generates an array [0, 1, 2, ..., 7]
+    try {
+      const allResults = await Promise.all(
+        assetIds.map(assetId => getIncentivePools(client, assetId, option, checkAddress))
+      );
+
+      const allPools = allResults.flat();
+      const activePools = allPools.filter(pool => pool.available.trim() != '0');
+
+      const summedRewards = activePools.reduce((acc, pool) => {
+        const assetId = pool.asset_id.toString();
+        const availableDecimal = (BigInt(pool.available) / BigInt(10 ** 27)).toString();
+        const availableFixed = (Number(availableDecimal) / 10 ** 9).toFixed(5); // Adjust for 5 decimal places
+
+        if (!acc[assetId]) {
+          acc[assetId] = { asset_id: assetId, funds: pool.funds, available: availableFixed };
+        } else {
+          acc[assetId].available = (parseFloat(acc[assetId].available) + parseFloat(availableFixed)).toFixed(5);
+        }
+
+        return acc;
+      }, {} as { [key: string]: { asset_id: string, funds: string, available: string } });
+
+      if (prettyPrint) {
+        const coinDictionary: { [key: string]: string } = {
+          '0': 'Sui',
+          '1': 'USDC',
+          '2': 'USDT',
+          '3': 'WETH',
+          '4': 'CETUS',
+          '5': 'vSui',
+          '6': 'haSui',
+          '7': 'NAVX',
+        };
+        console.log(checkAddress, ' available rewards:');
+        Object.keys(summedRewards).forEach(key => {
+          if (key == '5' || key == '7') {
+            console.log(`${coinDictionary[key]}: ${summedRewards[key].available} NAVX`);
+          } else {
+            console.log(`${coinDictionary[key]}: ${summedRewards[key].available} vSui`);
+          }
+        });
+      }
+
+      return summedRewards;
+    } catch (error) {
+      console.error('Failed to get available rewards:', error);
+      throw error;
+    }
+  }
+
+/**
+   * Claims all available rewards for the specified account.
+   * @returns PTB result
+   */
+export async function claimAllRewardsPTB(client: SuiClient, userToCheck: string) {
+    let txb = new Transaction();
+
+    const rewardsSupply: { [key: string]: Reward } = await getAvailableRewards(client, userToCheck, 1, false);
+    // Convert the rewards object to an array of its values
+    const rewardsArray: Reward[] = Object.values(rewardsSupply);
+    for (const reward of rewardsArray) {
+      await claimRewardFunction(txb, reward.funds, reward.asset_id, 1);
+    }
+
+    const rewardsBorrow: { [key: string]: Reward } = await getAvailableRewards(client, userToCheck, 3, false);
+    // Convert the rewards object to an array of its values
+    const rewardsBorrowArray: Reward[] = Object.values(rewardsBorrow);
+    for (const reward of rewardsBorrowArray) {
+      await claimRewardFunction(txb, reward.funds, reward.asset_id, 3);
+    }
+
+    return txb;
+  }
