+++ src/libs/PTB/index.ts
@@ -1,1 +1,1 @@
-export {depositCoin, depositCoinWithAccountCap, stakeTovSui, unstakeTovSui, withdrawCoin, withdrawCoinWithAccountCap, getHealthFactor, borrowCoin, repayDebt, returnMergedCoins, flashloan,repayFlashLoan, SignAndSubmitTXB, liquidateFunction, claimRewardFunction} from './commonFunctions';
+export {depositCoin, depositCoinWithAccountCap, stakeTovSui, unstakeTovSui, withdrawCoin, withdrawCoinWithAccountCap, getHealthFactor, borrowCoin, repayDebt, returnMergedCoins, flashloan,repayFlashLoan, SignAndSubmitTXB, liquidateFunction, claimRewardFunction, getIncentivePools, getAvailableRewards, claimAllRewardsPTB} from './commonFunctions';
